requirements.txt
instagrapi
cryptography
pyyaml
rich
moviepy